fix: update copyright year and upgrade jubilant dependency to v1.8
